Add test-suite for createMockedWindow
